updated project

+++ demo/src/main/java/com/example/demo/repository/DoctorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.Doctor;
 
 @Repository
-public interface DoctorRepository extends JpaRepository<Doctor, Long> {
+public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
 
 	Optional<Doctor> findByEmail(String email);
 }
